// frontend/types/patient/personalInformation.ts
export interface PersonalInformation {
    id: number;
    patient_id: number;
    birth_place: string;
    birth_date: string;
    gender: string;
    blood_type: string;
    maritial_status: string;
    religion: string;
    contact_number: string;
    email: string;
    id_type: string;
    id_number: string;
    employeer: string;
    education: string;
}